feat/#19/내 강의 페이지 css 구현, 사이드바 정리
