+++ .idea/workspace.xml
@@ -5,18 +5,21 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="c756325b-e896-40b1-a1e0-b5b3d636f0ca" name="Changes" comment="feat/#19/내 강의 페이지 구현">
-      <change afterPath="$PROJECT_DIR$/src/components/common/MyCourseSidebar.js" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/components/course/lists/DailyAttendance.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/components/items/MyCourseListItem.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/fonts/Pretendard-Black.ttf" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/fonts/Pretendard-Bold.ttf" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/fonts/Pretendard-Light.ttf" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/fonts/Pretendard-Medium.ttf" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/fonts/Pretendard-Regular.ttf" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/fonts/Pretendard-SemiBold.ttf" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/App.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/App.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/apis/MyCourseAPICalls.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/apis/MyCourseAPICalls.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/common/MyCourseSidebar.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/common/MyCourseSidebar.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/components/course/lists/InProgressList.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/course/lists/InProgressList.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/components/course/lists/LastCourseList.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/course/lists/LastCourseList.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/components/course/lists/MyCourseMain.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/course/lists/MyCourseMain.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/components/course/lists/MyCourseMain.js" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/components/course/lists/ScheduledCourseList.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/course/lists/ScheduledCourseList.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/components/modal/MyLectureDetailInfo.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/components/modal/MyLectureDetailInfo.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/course.css" beforeDir="false" afterPath="$PROJECT_DIR$/src/course.css" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/modules/MyCourseModule.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/modules/MyCourseModule.js" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -37,6 +40,7 @@
       </map>
     </option>
     <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+    <option name="RESET_MODE" value="HARD" />
   </component>
   <component name="MarkdownSettingsMigration">
     <option name="stateVersion" value="1" />
@@ -57,7 +61,7 @@
     "RunOnceActivity.ShowReadmeOnStart": "true",
     "WebServerToolWindowFactoryState": "false",
     "git-widget-placeholder": "feat/#19/my__lecture",
-    "last_opened_file_path": "C:/Lecture/13_restapi/lecture-source/chap03-comprehensice-react",
+    "last_opened_file_path": "C:/Lecture/final/hicoding-front/src/fonts",
     "node.js.detected.package.eslint": "true",
     "node.js.detected.package.tslint": "true",
     "node.js.selected.package.eslint": "(autodetect)",
@@ -72,6 +76,11 @@
     ]
   }
 }]]></component>
+  <component name="RecentsManager">
+    <key name="CopyFile.RECENT_KEYS">
+      <recent name="C:\Lecture\final\hicoding-front\src\fonts" />
+    </key>
+  </component>
   <component name="RunManager">
     <configuration name="npmStart" type="js.build_tools.npm">
       <package-json value="$PROJECT_DIR$/package.json" />
