+++ src/App.js
@@ -16,7 +16,6 @@ import {MEMBER_PATH} from "./apis/MemberAPICalls";
 import LectureMain from "./pages/lecture/LectureMain";
 import CourseMain from "./pages/course/CourseMain";
 import InProgressList from "./components/course/lists/InProgressList";
-import MyCourseMain from "./components/course/lists/MyCourseMain";
 import MyLectureDetailInfo from "./components/modal/MyLectureDetailInfo";
 import LastCourseList from "./components/course/lists/LastCourseList";
 import ScheduledCourseList from "./components/course/lists/ScheduledCourseList";
@@ -33,13 +32,14 @@ function App() {
                         <Route index element={<Main/>}/>
                         <Route path="lecture" element={<LectureMain/>}/>
                         <Route path="course" element={<CourseMain/>}/>
-                        <Route path="mylecture" element={<MyCourseMain/>}>
+                        <Route path="mylecture">
                             <Route path="inprogress" element={<InProgressList/>}/>
+                            <Route path="lastcourse" element={<LastCourseList/>}/>
+                            <Route path="scheduledcourse" element={<ScheduledCourseList/>}/>
                             <Route path=":cosCode" element={<MyLectureDetailInfo/>}/>
                         </Route>
                         <Route path="mylecture">
-                            <Route path="lastcourse" element={<LastCourseList/>}/>
-                            <Route path="scheduledcourse" element={<ScheduledCourseList/>}/>
+
                         </Route>
                         <Route path="attendance">
                             <Route path="day/:cosCode" element={<DailyAttendance/>}/>
